requirements.txt
python-dotenv
mysql-connector-python
cx_Oracle